requirements.txt
pyyaml
tensorflow==2.5.0
wikipedia
flask
flask-wtf
Pillow
